kvtenantccl: unskip TestTenantUpgradeInterlock
It does takes an enormous amount of time to run successfully, more
than the previous timeout allowed:

```
--- PASS: TestTenantUpgradeInterlock (316.36s)
    --- PASS: TestTenantUpgradeInterlock/lagging_binary_version (120.85s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_first_check_for_instances (9.02s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_fence_RPC (10.11s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_fence_write_to_settings_table (20.27s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_second_check_of_instances (20.20s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_migration (20.45s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_version_bump_RPC (20.28s)
        --- PASS: TestTenantUpgradeInterlock/lagging_binary_version/pause_after_write_to_settings_table (20.51s)
    --- PASS: TestTenantUpgradeInterlock/current_binary_version (195.51s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_first_check_for_instances (20.00s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_fence_RPC (30.23s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_fence_write_to_settings_table (29.02s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_second_check_of_instances (29.14s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_migration (29.00s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_version_bump_RPC (29.03s)
        --- PASS: TestTenantUpgradeInterlock/current_binary_version/pause_after_write_to_settings_table (29.08s)
PASS
INFO: Elapsed time: 318.059s, Critical Path: 317.36s
INFO: 2 processes: 1 internal, 1 processwrapper-sandbox.
//pkg/ccl/kvccl/kvtenantccl:kvtenantccl_test                             PASSED in 317.2s
```

(and this is on a faster machine than in CI)

Release note: None
