+++ pkg/BUILD.bazel
@@ -40,6 +40,8 @@ ALL_TESTS = [
     "//pkg/ccl/jobsccl/jobsprotectedtsccl:jobsprotectedtsccl_test",
     "//pkg/ccl/jwtauthccl:jwtauthccl_test",
     "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl_test",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeccl:upgradeccl_test",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl:upgradeinterlockccl_test",
     "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
@@ -775,6 +777,8 @@ GO_TARGETS = [
     "//pkg/ccl/jwtauthccl:jwtauthccl_test",
     "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl",
     "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl_test",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeccl:upgradeccl_test",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl:upgradeinterlockccl_test",
     "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl",
     "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl_test",
     "//pkg/ccl/kvccl:kvccl",
@@ -2417,6 +2421,8 @@ GET_X_DATA_TARGETS = [
     "//pkg/ccl/kvccl:get_x_data",
     "//pkg/ccl/kvccl/kvfollowerreadsccl:get_x_data",
     "//pkg/ccl/kvccl/kvtenantccl:get_x_data",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeccl:get_x_data",
+    "//pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl:get_x_data",
     "//pkg/ccl/logictestccl:get_x_data",
     "//pkg/ccl/logictestccl/tests/3node-tenant:get_x_data",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:get_x_data",
