+++ pkg/ccl/kvccl/kvtenantccl/BUILD.bazel
@@ -54,7 +54,6 @@ go_test(
         "tenant_range_lookup_test.go",
         "tenant_scan_range_descriptors_test.go",
         "tenant_trace_test.go",
-        "tenant_upgrade_test.go",
     ],
     args = ["-test.timeout=895s"],
     embed = [":kvtenantccl"],
@@ -63,10 +62,8 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/clusterversion",
         "//pkg/config",
         "//pkg/gossip",
-        "//pkg/jobs",
         "//pkg/keys",
         "//pkg/kv/kvclient/kvcoord",
         "//pkg/kv/kvclient/kvtenant",
@@ -80,24 +77,13 @@ go_test(
         "//pkg/security",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
-        "//pkg/security/username",
         "//pkg/server",
         "//pkg/settings",
-        "//pkg/settings/cluster",
-        "//pkg/spanconfig",
         "//pkg/sql",
-        "//pkg/sql/catalog/lease",
-        "//pkg/sql/sessiondatapb",
-        "//pkg/sql/sqlinstance/instancestorage",
-        "//pkg/sql/sqlliveness/slinstance",
-        "//pkg/sql/stats",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
-        "//pkg/testutils/skip",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
-        "//pkg/upgrade",
-        "//pkg/upgrade/upgradebase",
         "//pkg/util",
         "//pkg/util/grpcutil",
         "//pkg/util/hlc",
